Moved Drug mocklist to FirstFragment.kt and created the Recyclerview there
