+++ app/src/main/java/com/saotome/imed_app/MainActivity.kt
@@ -10,14 +10,11 @@ import androidx.navigation.ui.setupActionBarWithNavController
 import android.view.Menu
 import android.view.MenuItem
 import com.saotome.imed_app.databinding.ActivityMainBinding
-import com.saotome.imed_app.model.Drug
 
 class MainActivity : AppCompatActivity() {
 
     private lateinit var appBarConfiguration: AppBarConfiguration
     private lateinit var binding: ActivityMainBinding
-
-    private val drugList = arrayListOf<Drug>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -35,14 +32,6 @@ class MainActivity : AppCompatActivity() {
             Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
                 .setAction("Action", null).show()
         }
-
-        loadDrugMockList()
-    }
-
-    private fun loadDrugMockList() {
-        drugList.add(0, Drug(0, "Unak", "1 capsula", "20:00", "24:00"))
-        drugList.add(0, Drug(0, "Glifage", "1 capsula", "20:00", "24:00"))
-        drugList.add(0, Drug(0, "Indapen", "1 capsula", "10:00", "48:00"))
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
